refact: remove duplicate key functions
