+++ pkg/common/configuration/config.go
@@ -1,15 +1,12 @@
 package configuration
 
 import (
-	"crypto/x509"
-	"encoding/pem"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	"time"
 
 	"github.com/kelseyhightower/envconfig"
-	"gopkg.in/square/go-jose.v2"
+	"github.com/stone-co/webhook-consumer/pkg/common/keys"
 )
 
 // Config defines the service configuration
@@ -39,7 +36,7 @@ func LoadConfig() (*Config, error) {
 		return nil, fmt.Errorf("reading file %s: %v", config.PrivateKeyPath, err)
 	}
 
-	config.PrivateKey, err = LoadPrivateKey(keyBytes)
+	config.PrivateKey, err = keys.LoadPrivateKey(keyBytes)
 	if err != nil {
 		return nil, fmt.Errorf("unable to read private key: %v", err)
 	}
@@ -49,89 +46,10 @@ func LoadConfig() (*Config, error) {
 		return nil, fmt.Errorf("reading file %s: %v", config.PublicKeyPath, err)
 	}
 
-	config.VerificationKey, err = LoadPublicKey(keyBytes)
+	config.VerificationKey, err = keys.LoadPublicKey(keyBytes)
 	if err != nil {
 		return nil, fmt.Errorf("unable to read public key: %v", err)
 	}
 
 	return &config, nil
-}
-
-// LoadPrivateKey loads a private key from PEM/DER/JWK-encoded data.
-func LoadPrivateKey(data []byte) (interface{}, error) {
-	input := data
-
-	block, _ := pem.Decode(data)
-	if block != nil {
-		input = block.Bytes
-	}
-
-	var priv interface{}
-	priv, err0 := x509.ParsePKCS1PrivateKey(input)
-	if err0 == nil {
-		return priv, nil
-	}
-
-	priv, err1 := x509.ParsePKCS8PrivateKey(input)
-	if err1 == nil {
-		return priv, nil
-	}
-
-	priv, err2 := x509.ParseECPrivateKey(input)
-	if err2 == nil {
-		return priv, nil
-	}
-
-	jwk, err3 := LoadJSONWebKey(input, false)
-	if err3 == nil {
-		return jwk, nil
-	}
-
-	return nil, fmt.Errorf("square/go-jose: parse error, got '%s', '%s', '%s' and '%s'", err0, err1, err2, err3)
-}
-
-func LoadJSONWebKey(json []byte, pub bool) (*jose.JSONWebKey, error) {
-	var jwk jose.JSONWebKey
-	err := jwk.UnmarshalJSON(json)
-	if err != nil {
-		return nil, err
-	}
-
-	if !jwk.Valid() {
-		return nil, errors.New("invalid JWK key")
-	}
-
-	if jwk.IsPublic() != pub {
-		return nil, errors.New("priv/pub JWK key mismatch")
-	}
-
-	return &jwk, nil
-}
-
-// LoadPublicKey loads a public key from PEM/DER/JWK-encoded data.
-func LoadPublicKey(data []byte) (interface{}, error) {
-	input := data
-
-	block, _ := pem.Decode(data)
-	if block != nil {
-		input = block.Bytes
-	}
-
-	// Try to load SubjectPublicKeyInfo
-	pub, err0 := x509.ParsePKIXPublicKey(input)
-	if err0 == nil {
-		return pub, nil
-	}
-
-	cert, err1 := x509.ParseCertificate(input)
-	if err1 == nil {
-		return cert.PublicKey, nil
-	}
-
-	jwk, err2 := LoadJSONWebKey(data, true)
-	if err2 == nil {
-		return jwk, nil
-	}
-
-	return nil, fmt.Errorf("square/go-jose: parse error, got '%s', '%s' and '%s'", err0, err1, err2)
 }
